Clearer list of activationHooks
The list of known activation hooks was previously confusing because the list had only one item, and when rendered showed up as a bullet point in the same level list of `package.json` properties.

The only hook I could find any mention of was the one that was already there, so I assume it's the only one. Changed the wording to indicate that.

Also made the use of e.g. clearer w/ parentheses.