Break commands in two lines at Linux instructions.
When the commands are put together in the same line of command at bash, the admin password and curl prompts are mixed at bash, which may lead to a confusing experience for the Atom's users during the installation process (the user may not see at first the sudo prompt and keep waiting for the curl result). Thanks for considering this little change.