Update atom-selections.md
Changed key combination sequence by moving ```Shift``` to the front of sequence. For the reader this emphasise ```Shift``` as concept of selection. Psychologically it is intuitive to start pressing ```Shift``` to select by cursor movement. This also makes the cursor movement key combination more obvious for the analytical user. Exception is ```Ctrl-Shift-W``` which is unchanged because it is special.