Fix "stdout maxBuffer exceeded" error when building site
Prior to this change, the the 'nanoc:compile' gulp task attempted to 
store the complete output of `nanoc:compile` in a string, which resulted 
in a "stdout maxBuffer exceeded" error. By using `spawn` instead of 
`exec`, we can pipe the output of `bundle exec nanoc compile` to STDIO 
and thus avoid trying to fit it all in a string.

Example build failure: 
https://travis-ci.org/atom/flight-manual.atom.io/builds/550937870#L3656