Don't hard-wrap when encountering newlines
With the introduction of kramdown in 73b0e94, we started interpreting
newline characters as hard wraps, which was causing paragraphs to not
take advantage of all the available screen real estate. Documentation is
usually hard-wrapped for readability purposes when navigating source
code, but those concerns don't apply when viewing the same documentation
in a web browser.

This commit fixes the visual glitch illustrated above by supplying
`hard_wrap: false` to  kramdown-gfm. I verified that this doesn't affect
the rendering of other markdown constructs such as code block syntax
highlighting.