Update find-and-replace.md
Previously, it was unclear to me how to specify a single folder to search in. It was not obvious that you could actually type the name o a folder in that box at the bottom (under 'Replace in Project') to refine your search. The screenshot should help others see this option more clearly.